move constant SHUTDOWN_PHASER from ShutdownHookCallbacks to CommonConstants
