+++ dubbo-common/src/main/java/org/apache/dubbo/common/lang/ShutdownHookCallbacks.java
@@ -21,9 +21,9 @@ import org.apache.dubbo.common.extension.ExtensionLoader;
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.concurrent.Phaser;
 
 import static java.util.Collections.sort;
+import static org.apache.dubbo.common.constants.CommonConstants.SHUTDOWN_PHASER;
 import static org.apache.dubbo.common.function.ThrowableAction.execute;
 
 /**
@@ -37,14 +37,7 @@ public class ShutdownHookCallbacks {
 
     private final List<ShutdownHookCallback> callbacks = new LinkedList<>();
 
-    /**
-     * This variable is used to make sure that the asynchronous thread can be executed correctly
-     * in the case of graceful shutdown.
-     * <br/><br/>
-     * When you have a thread that must be run asynchronously when shutting down.
-     * You can register the phaser before the thread run, and unregister this when the thread is finished.
-     */
-    public static final Phaser SHUTDOWN_PHASER = new Phaser(1);
+
 
     ShutdownHookCallbacks() {
         loadCallbacks();
